LayoutTests/imported/w3c/web-platform-tests/css/css-values/viewport-units-scrollbars-root-properties-001-expected.txt

Measure to left

FAIL vw has had the overflow-y scrollbar width removed assert_less_than: expected a number less than 80 but got 80
